feat: Migrate `Category.DisplayOrder` to decimal type and enhance frontend category reordering with moved item ID.

+++ ProductCatalogSystem.Server/Domain/Category.cs
@@ -12,7 +12,7 @@ public sealed class Category : AuditableEntity
 
     public CategoryStatus Status { get; set; }
 
-    public int DisplayOrder { get; set; }
+    public decimal DisplayOrder { get; set; }
 
     public byte[] RowVersion { get; set; } = [];
